CMR cleanup for LinodeDiskActionMenu
